backend/requirements.txt
openai
SpeechRecognition
sentence_transformers
psycopg2_binary